develop: 구조 변경

+++ src/main/java/com/sollertia/habit/domain/habit/dto/HabitDtoImpl.java
@@ -1,12 +1,7 @@
 package com.sollertia.habit.domain.habit.dto;
 
-import com.sollertia.habit.domain.habit.enums.Day;
 import com.sollertia.habit.domain.category.enums.Category;
-import com.sollertia.habit.domain.team.Team;
-import com.sollertia.habit.domain.user.User;
 import lombok.*;
-
-import java.util.List;
 
 @Getter
 @Setter
@@ -21,7 +16,7 @@ public class HabitDtoImpl {
     private String durationStart;
     private String durationEnd;
     private Long categoryId;
-    private Long count;
+    private int count;
     private Long goalTime = 0l;
     private String practiseDays;
     private Category categoryType;
